Move MediaRepository to media_repository module
